+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.0.0] - 2021-05-15
+### Added
+- TypeScript support
+### Fixed
+- cleaning up stack traces for newer Node versions
+### Changed
+- named exports instead of default
+- default options are now changed via function instead of overriding property of createError
+
 ## [3.0.1] - 2020-03-13
 ### Fixed
 - CommonJS fallback